Implement data URIs.
In order to further reduce the number of HTTP requests, sprites may now
be "inlined" in a Sass file using the data URI system. To enable this,
use :data_uri as the sprite path in your .polymer file.

Polymer adds the data URI to the Sass file by using a selector (such as
".my_sprite_data"), and then uses Sass' @extend feature so that the data
is only added to the CSS file once. Your own Sass files do not need to
be modified to make use of this feature; the @polymer mixin suffices.