Use ChunkyPNG in specs and features.
This begins the switch to ChunkyPNG for all operations; RMagick isn't
all that portable and, despite being fast, can be a pain for
inexperienced users to install.

ChunkyPNG means that we'll only be able to use PNG sources and sprites,
but since PNGs are often the best format for such images, this isn't
much of a problem.